+++ demo/code.ts
@@ -2,6 +2,9 @@ import { MyA, MyB, MyC, MyD, MyE, MyF } from "./classes";
 
 const a = new MyA();
 const b = new MyB();
+
+function /* def */ field() {}
+
 const c = new MyC();
 const d = new MyD();
 const e = new MyE();