+++ demo/demo.js
@@ -1,0 +1,34 @@
+main();
+
+async function main() {
+	bar();
+	await asyncFn1();
+}
+
+async function asyncFn1() {
+	console.log("asyncFn1");
+	await asyncFn2();
+	await asyncFn2();
+}
+
+async function asyncFn2() {
+	await new Promise((res) => setTimeout(res, 10));
+	console.log("asyncFn2");
+}
+
+function bar() {
+	console.log("foo");
+
+	a(10);
+}
+
+function a(i) {
+	if (i === 0) {
+		return 0;
+	}
+	return b(i - 1);
+}
+
+function b(i) {
+	return a(i - 1);
+}