+++ demo/simple-demo.ts
@@ -1,0 +1,10 @@
+// === New Way - Multi Rename ===
+
+export class MyA {
+	myField1 = 1;
+	myField2 = 2;
+}
+
+const m = new MyA();
+console.log(m.myField1);
+console.log(m.myField2);